🧑‍💻 Generate floats instead of doubles
Everything I'm passing these values too (e.g. yoga) wants floats anyway. Let's get rid of conversion warnings

+++ src/xaml-to-fui-statictheme/GetResources.cpp
@@ -59,7 +59,7 @@ void GetResource(
     return;
   }
   if (tagType == "x:Double") {
-    GetNumber(back, it, "double");
+    GetNumber(back, it, "float");
     return;
   }
   if (tagType == "x:Int32") {
